feat: move tools to features+add link in main.scss
