+++ src/assets/styles/main.scss
@@ -7,4 +7,4 @@
 @import 'components/buttons', 'components/logo', 'components/titles', 'components/inputs', 'components/text',
     'components/wrappers', 'components/forms';
 
-@import '../../layouts/header/header';
+@import '../../layouts/header/header', '../../layouts/user-page/user-page;
